Refactor: 기존 user service의 토큰 업데이트 기능 분리 #43
- 기존 user_service의 update_user_fcm_token 삭제 및 fcm_service로 이동
- 기존 user_handler 삭제 및 fcm_handler로 이동
